+++ src/main/java/roomescape/domain/theme/ThemeName.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.domain.theme;
 
 import roomescape.domain.exception.InvalidDomainObjectException;
